requirements.txt
fastapi
uvicorn
pytest
pytest-cov
httpx
PyJWT